# catalog/lease: remove expiry based leasing code