--- a/pkg/settings/registry.go
+++ b/pkg/settings/registry.go
@@ -256,6 +256,9 @@ var retiredSettings = map[InternalKey]struct{}{
 	"jobs.execution_errors.max_entry_size":                       {},
 	"sql.metrics.statement_details.plan_collection.enabled":      {},
 	"sql.metrics.statement_details.plan_collection.period":       {},
+
+	// removed as of 25.2
+	"sql.catalog.experimental_use_session_based_leasing": {},
 }
 
 // grandfatheredDefaultSettings is the list of "grandfathered" existing sql.defaults