--- a/pkg/sql/catalog/lease/BUILD.bazel
+++ b/pkg/sql/catalog/lease/BUILD.bazel
@@ -1,5 +1,4 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-load("//build:STRINGER.bzl", "stringer")
 
 go_library(
     name = "lease",
@@ -14,7 +13,6 @@ go_library(
         "name_cache.go",
         "storage.go",
         "testutils.go",
-        ":gen-lease-mode-stringer",  # keep
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/sql/catalog/lease",
     visibility = ["//visibility:public"],
@@ -70,12 +68,6 @@ go_library(
         "@com_github_cockroachdb_logtags//:logtags",
         "@com_github_cockroachdb_redact//:redact",
     ],
-)
-
-stringer(
-    name = "gen-lease-mode-stringer",
-    src = "lease.go",
-    typ = "SessionBasedLeasingMode",
 )
 
 go_test(