--- a/pkg/sql/catalog/lease/descriptor_version_state.go
+++ b/pkg/sql/catalog/lease/descriptor_version_state.go
@@ -7,7 +7,6 @@ package lease
 
 import (
 	"context"
-	"time"
 
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog"
 	"github.com/cockroachdb/cockroach/pkg/sql/catalog/descpb"
@@ -16,7 +15,6 @@ import (
 	"github.com/cockroachdb/cockroach/pkg/util/hlc"
 	"github.com/cockroachdb/cockroach/pkg/util/log"
 	"github.com/cockroachdb/cockroach/pkg/util/syncutil"
-	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
 	"github.com/cockroachdb/redact"
 )
 
@@ -140,26 +138,13 @@ func (s *descriptorVersionState) incRefCountLocked(ctx context.Context, expensiv
 }
 
 func (s *descriptorVersionState) getExpirationLocked(ctx context.Context) hlc.Timestamp {
-	// A descriptor version state can now potentially contain two different types
-	// of expiration:
-	// 1) Fixed expirations, which will be based on some timestamp in the future,
-	//   that will need to be renewed to keep a descriptor as "active"
-	// 2) Session-based expirations, which say that a descriptor is in use,
-	//    as long as the sqlliveness exists for it.
-	// We are going to pick the longest possible leases between these two options,
-	// assuming that session-based leases are being enforced. Session-based leases
-	// will only be enforced once the Drain leasing mode is reached, which will stop
-	// allowing fixed expiration leases from renewing (i.e. those leases will
-	// eventually be *drained*.
-	expiration := s.mu.expiration
-	if s.mu.session != nil &&
-		s.t.m.sessionBasedLeasingModeAtLeast(ctx, SessionBasedDrain) {
-		sessionExpiry := s.mu.session.Expiration()
-		if expiration.Less(sessionExpiry) {
-			expiration = sessionExpiry
-		}
+	// If an expiration is set then this descriptor is a stale version,
+	// and will be eventually removed.
+	if !s.mu.expiration.IsEmpty() {
+		return s.mu.expiration
 	}
-	return expiration
+	// Otherwise, the expiration is tied to sqlliveness.
+	return s.mu.session.Expiration()
 }
 
 func (s *descriptorVersionState) getExpiration(ctx context.Context) hlc.Timestamp {
@@ -167,23 +152,4 @@ func (s *descriptorVersionState) getExpiration(ctx context.Context) hlc.Timestam
 	defer s.mu.Unlock()
 
 	return s.getExpirationLocked(ctx)
-}
-
-// getStoredLease returns a copy of the stored lease.
-func (s *descriptorVersionState) getStoredLease() *storedLease {
-	s.mu.Lock()
-	defer s.mu.Unlock()
-	if s.mu.lease == nil {
-		return nil
-	}
-	leaseCopy := *s.mu.lease
-	return &leaseCopy
-}
-
-// The lease expiration stored in the database is of a different type.
-// We've decided that it's too much work to change the type to
-// hlc.Timestamp, so we're using this method to give us the stored
-// type: tree.DTimestamp.
-func storedLeaseExpiration(expiration hlc.Timestamp) tree.DTimestamp {
-	return tree.DTimestamp{Time: timeutil.Unix(0, expiration.WallTime).Round(time.Microsecond)}
 }