--- a/pkg/sql/catalog/lease/kv_writer.go
+++ b/pkg/sql/catalog/lease/kv_writer.go
@@ -27,9 +27,6 @@ type kvWriter struct {
 	// Used to write leases that have will no longer have an expiration,
 	// but have their lifetime tied to a sqlliveness session.
 	sessionBasedWriter bootstrap.KVWriter
-	// Used to write leases that would use an expiration time
-	// previously.
-	expiryBasedWriter bootstrap.KVWriter
 
 	settingsWatcher    *settingswatcher.SettingsWatcher
 	sessionBasedReader sessionBasedLeasingModeReader
@@ -45,7 +42,6 @@ func newKVWriter(
 	return &kvWriter{
 		db:                 db,
 		sessionBasedWriter: bootstrap.MakeKVWriter(codec, leaseTableWithID(id, systemschema.LeaseTable())),
-		expiryBasedWriter:  bootstrap.MakeKVWriter(codec, leaseTableWithID(id, systemschema.LeaseTable_V23_2())),
 		settingsWatcher:    settingsWatcher,
 		sessionBasedReader: sessionModeReader,
 	}
@@ -64,23 +60,8 @@ func leaseTableWithID(id descpb.ID, table systemschema.SystemTable) catalog.Tabl
 
 func (w *kvWriter) insertLease(ctx context.Context, txn *kv.Txn, l leaseFields) error {
 	if err := w.do(ctx, txn, l, func(b *kv.Batch) error {
-		// We support writing both session based and expiry based leases within
-		// the KV writer. To be able to support a migration between the two types
-		// of writer will in some cases need to be able to write both types of leases.
-		// As a result based on our currently active mode, determine which types
-		// of leases should be written. The scenarios we support are:
-		// 1) Session Based Off => Only expiry based leases are written.
-		// 2) Dual-Write => Both session and expiry based leases will be written.
-		// 3) Session Only => Only session based leases will get written.
-		if w.sessionBasedReader.sessionBasedLeasingModeAtLeast(ctx, SessionBasedDualWrite) &&
-			l.sessionID != nil {
+		if l.sessionID != nil {
 			err := w.sessionBasedWriter.Insert(ctx, b, false /*kvTrace*/, leaseAsSessionBasedDatum(l)...)
-			if err != nil {
-				return err
-			}
-		}
-		if !w.sessionBasedReader.sessionBasedLeasingModeAtLeast(ctx, SessionBasedOnly) {
-			err := w.expiryBasedWriter.Insert(ctx, b, false /*kvTrace */, leaseAsRbrDatum(l)...)
 			if err != nil {
 				return err
 			}
@@ -94,23 +75,8 @@ func (w *kvWriter) insertLease(ctx context.Context, txn *kv.Txn, l leaseFields) 
 
 func (w *kvWriter) deleteLease(ctx context.Context, txn *kv.Txn, l leaseFields) error {
 	if err := w.do(ctx, txn, l, func(b *kv.Batch) error {
-		// We support deleting both session based and expiry based leases within
-		// the KV writer. To be able to support a migration between the two types
-		// of writer will in some cases need to be able to delete both types of leases.
-		// As a result based on our currently active mode, determine which types
-		// of leases should be deleted. The scenarios we support are:
-		// 1) Session Based Off => Only expiry based leases are deleted.
-		// 2) Dual-Write => Both session and expiry based leases will be deleted.
-		// 3) Session Only => Only session based leases will get deleted.
-		if w.sessionBasedReader.sessionBasedLeasingModeAtLeast(ctx, SessionBasedDualWrite) &&
-			l.sessionID != nil {
+		if l.sessionID != nil {
 			err := w.sessionBasedWriter.Delete(ctx, b, false /*kvTrace*/, leaseAsSessionBasedDatum(l)...)
-			if err != nil {
-				return err
-			}
-		}
-		if !w.sessionBasedReader.sessionBasedLeasingModeAtLeast(ctx, SessionBasedOnly) {
-			err := w.expiryBasedWriter.Delete(ctx, b, false /*kvTrace */, leaseAsRbrDatum(l)...)
 			if err != nil {
 				return err
 			}
@@ -150,15 +116,4 @@ func leaseAsSessionBasedDatum(l leaseFields) []tree.Datum {
 		tree.NewDBytes(tree.DBytes(l.sessionID)),
 		tree.NewDBytes(tree.DBytes(l.regionPrefix)),
 	}
-}
-
-func leaseAsRbrDatum(l leaseFields) []tree.Datum {
-	return []tree.Datum{
-		tree.NewDInt(tree.DInt(l.descID)),
-		tree.NewDInt(tree.DInt(l.version)),
-		tree.NewDInt(tree.DInt(l.instanceID)),
-		&l.expiration,
-		tree.NewDBytes(tree.DBytes(l.regionPrefix)),
-	}
-
 }