Previously, we had a lot of complexity inside the lease manager code to help support migration from expiry based leasing in code. As a result we needed to support complex logic to support both variants of leasing. This patch aims to reduce the complexity of the code by doing the following:

1. Remove support for leasing modes outside of session based leasing
2. Remove the leasing mode setting

Running the Sysbench repeated run we see a small improvement:
```
With changes:

Feb 10 14:02:37 faizan-perf-0004 sak-repeated-run.sh[20693]:     transactions:                        722299 (1203.76 per sec.)
Feb 10 14:02:37 faizan-perf-0004 sak-repeated-run.sh[20693]:     queries:                             14446005 (24075.14 per sec.)
Feb 10 14:22:27 faizan-perf-0004 sak-repeated-run.sh[21520]:     transactions:                        715476 (1192.36 per sec.)
Feb 10 14:22:27 faizan-perf-0004 sak-repeated-run.sh[21520]:     queries:                             14309520 (23847.25 per sec.)
Feb 10 14:41:50 faizan-perf-0004 sak-repeated-run.sh[22204]:     transactions:                        699994 (1166.59 per sec.)
Feb 10 14:41:50 faizan-perf-0004 sak-repeated-run.sh[22204]:     queries:                             13999911 (23331.91 per sec.)
Feb 10 15:01:00 faizan-perf-0004 sak-repeated-run.sh[22986]:     transactions:                        706105 (1176.72 per sec.)
Feb 10 15:01:00 faizan-perf-0004 sak-repeated-run.sh[22986]:     queries:                             14122100 (23534.33 per sec.)
Feb 10 15:20:02 faizan-perf-0004 sak-repeated-run.sh[23681]:     transactions:                        696832 (1161.31 per sec.)
Feb 10 15:20:02 faizan-perf-0004 sak-repeated-run.sh[23681]:     queries:                             13936640 (23226.24 per sec.

Mean TPS: 1180.15

Without changes:
Feb 10 18:07:27 faizan-perf-0004 sak-repeated-run.sh[26236]:     transactions:                        700825 (1167.97 per sec.)
Feb 10 18:07:27 faizan-perf-0004 sak-repeated-run.sh[26236]:     queries:                             14016500 (23359.40 per sec.)
Feb 10 18:26:26 faizan-perf-0004 sak-repeated-run.sh[26935]:     transactions:                        689313 (1148.72 per sec.)
Feb 10 18:26:26 faizan-perf-0004 sak-repeated-run.sh[26935]:     queries:                             13786272 (22974.44 per sec.)
Feb 10 18:45:14 faizan-perf-0004 sak-repeated-run.sh[27640]:     transactions:                        711582 (1185.90 per sec.)
Feb 10 18:45:14 faizan-perf-0004 sak-repeated-run.sh[27640]:     queries:                             14231640 (23717.95 per sec.)
Feb 10 19:04:09 faizan-perf-0004 sak-repeated-run.sh[28386]:     transactions:                        684539 (1140.84 per sec.)
Feb 10 19:04:09 faizan-perf-0004 sak-repeated-run.sh[28386]:     queries:                             13690820 (22816.79 per sec.)
Feb 10 19:23:20 faizan-perf-0004 sak-repeated-run.sh[29080]:     transactions:                        705886 (1176.35 per sec.)
Feb 10 19:23:20 faizan-perf-0004 sak-repeated-run.sh[29080]:     queries:                             14117748 (23527.00 per sec.)

Mean TPS: 1163.96
```

Benchmark Workload
```
name                                     old time/op    new time/op    delta
Sysbench/SQL/3node/oltp_read_only-10       3.03ms ±19%    2.76ms ±10%  -8.89%  (p=0.029 n=10+10)
Sysbench/SQL/3node/oltp_write_only-10      2.31ms ± 4%    2.25ms ± 6%    ~     (p=0.165 n=10+10)
Sysbench/SQL/3node/oltp_read_write-10      5.44ms ± 7%    5.38ms ± 5%    ~     (p=0.497 n=9+10)
Sysbench/SQL/3node/oltp_point_select-10     172µs ± 8%     170µs ± 4%    ~     (p=0.796 n=9+9)
Sysbench/SQL/3node/oltp_begin_commit-10     134µs ± 5%     132µs ± 7%    ~     (p=0.258 n=9+9)

name                                     old alloc/op   new alloc/op   delta
Sysbench/SQL/3node/oltp_read_only-10       1.13MB ± 2%    1.13MB ± 0%    ~     (p=0.573 n=10+8)
Sysbench/SQL/3node/oltp_write_only-10       902kB ± 1%     903kB ± 2%    ~     (p=0.796 n=10+10)
Sysbench/SQL/3node/oltp_read_write-10      2.07MB ± 0%    2.07MB ± 1%    ~     (p=0.605 n=9+9)
Sysbench/SQL/3node/oltp_point_select-10    31.4kB ± 2%    31.5kB ± 2%    ~     (p=0.353 n=10+10)
Sysbench/SQL/3node/oltp_begin_commit-10    11.4kB ± 4%    11.2kB ± 4%    ~     (p=0.239 n=10+10)

name                                     old allocs/op  new allocs/op  delta
Sysbench/SQL/3node/oltp_read_only-10        4.32k ± 4%     4.27k ± 0%    ~     (p=0.588 n=10+8)
Sysbench/SQL/3node/oltp_write_only-10       5.71k ± 1%     5.71k ± 1%    ~     (p=0.986 n=10+10)
Sysbench/SQL/3node/oltp_read_write-10       10.3k ± 1%     10.3k ± 1%    ~     (p=0.725 n=10+10)
Sysbench/SQL/3node/oltp_point_select-10       237 ± 2%       238 ± 2%    ~     (p=0.154 n=10+10)
Sysbench/SQL/3node/oltp_begin_commit-10      71.3 ± 5%      70.4 ± 4%    ~     (p=0.158 n=10+10)
```

Fixes: #133177
Fixes: #141032