@rafiss 2025-02-12
> It's somewhat related to your change, but I guess we could remove code that handled the old system.lease format. It's probably not a big perf win, so it's fine to leave out of this PR.

+1. Can a followup PR also remove the old lease table definition? (Or this one, if it doesn't add too much to the diff)

https://github.com/cockroachdb/cockroach/blob/33e120c6572ed5bc38d07a4c4e7f0cd46257deac/pkg/sql/catalog/systemschema/system.go#L145